% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_functions.R
\name{check_lubridate}
\alias{check_lubridate}
\title{Check if lubridate is installed}
\usage{
check_lubridate()
}
\value{
This function installs and loads the lubridate package in case is has not been downloaded and loaded
}
\description{
This function installs and loads the lubridate package in case is has not been downloaded and loaded
}
\examples{
\dontrun{
check_lubridate()
}

}
