% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA Significant Earthquake Database data}
\usage{
eq_clean_data(dataset)
}
\arguments{
\item{dataset}{Data set from the NOAA Significante Earthquakte Database}
}
\value{
This function returns a cleaned data set of the NOAA Significante Earthquakte Database containing a date column
}
\description{
This function takes the NOAA Significant Earthquake Database data as input and gives back a cleaned data set as a tibble.
}
\examples{
\dontrun{
eq_clean_data(NOAA_dataset)
}

}
