% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_maps.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Building eq_create_label}
\usage{
eq_create_label(dataframe)
}
\arguments{
\item{dataframe}{A data frame containing relevant information about earthquakes}
}
\description{
This function creates the annotation text for an HTML label of the leaflet map.
}
\examples{
\dontrun{
noaa_read("NOAA Significant Earthquake Database.txt") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")
}

}
