% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_maps.R
\name{eq_map}
\alias{eq_map}
\title{Building eq_map}
\usage{
eq_map(dataframe, annot_col)
}
\arguments{
\item{dataframe}{A data frame containing relevant information about earthquakes}

\item{annot_col}{Annotation data stored in the data frame}
}
\description{
This function creates a map of the earthquakes in a data set and allows for annotation of the data for instance for each point representing
an earthquake.
}
\examples{
\dontrun{
noaa_data \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")
}

}
