% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_geoms.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Building geom_timeline}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  size = size,
  ...
)
}
\arguments{
\item{size}{size of the points (optional parameter)}

\item{x}{Varaibel displayed on the x-axis, i.e. Date}

\item{alpha}{transparancy of the points (optional parameter)}

\item{color}{color of the points (optional parameter)}
}
\description{
Building geom_timeline
}
\examples{
\dontrun{
ggplot(data = usa_data) +
geom_timeline(aes(x = DATE, color = DEATHS, size=EQ_PRIMARY), alpha=0.6)
}

}
