% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_geoms.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Creating a new class "GeomTimeline"}
\format{An object of class \code{GeomTimeline} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeline
}
\description{
This geom object uses ggplot2's standard point geom, i.e. GeomPoint and will be used to build the geom_timeline geom.
}
\keyword{datasets}
