% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_geoms.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Creating a new class "GeomTimelineLabel"}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimelineLabel
}
\description{
This geom object uses ggplot2's standard geom, i.e. Geom and grid's segmentsGrob and textGrob to build the geom_timeline_label geom.
}
\keyword{datasets}
