% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_functions.R
\name{noaa_read}
\alias{noaa_read}
\title{NOAA Significant Earthquake Database data import}
\usage{
noaa_read(filename)
}
\arguments{
\item{filename}{A character string containing the path and the file name of a txt.-file.}
}
\value{
This function returns the imported file as a tibble data frame.
}
\description{
This function takes the NOAA Significant Earthquake Database txt.-files and imports the file in case it is available and returns a data frame
as a tibble.
Messages caused by the import are suppressed. In case the file does not exist, the program stops and prints
out that the file does not exist.
}
\note{
The function stops if the file name provided does not exist.
}
\examples{
\dontrun{
noaa_read("NOAA Significant Earthquake Database.txt")
}

}
