% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_geoms.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Building geom_timeline}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  size = size,
  ...
)
}
\arguments{
\item{mapping}{Internal parameter needed for the geom}

\item{data}{Internal parameter needed for the geom}

\item{stat}{Internal parameter needed for the geom}

\item{position}{Internal parameter needed for the geom}

\item{show.legend}{Internal parameter needed for the geom}

\item{na.rm}{Internal parameter needed for the geom}

\item{inherit.aes}{Internal parameter needed for the geom}

\item{size}{Internal parameter needed for the geom}

\item{...}{Internal parameter needed for the geom}
}
\description{
Building geom_timeline
}
\examples{
\dontrun{
noaa_read("NOAA Significant Earthquake Database.txt") \%>\%
eq_clean_data() \%>\%
filter(str_trim(COUNTRY) \%in\% c("USA") & DATE \%in\% c(ymd("2000-01-01"):ymd("2017-01-01"))) \%>\%
ggplot() +
geom_timeline(aes(x = DATE, color = DEATHS, size=EQ_PRIMARY), alpha=0.6)
}

}
