% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_geoms.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Building geom_timeline_label}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{x}{Variable displayed on the x-axis, i.e. Date}

\item{label}{Location name}

\item{size}{Variable representing the size (magnitude) of the earthquake, i.e. EQ_PRIMARY}

\item{n_max}{Amount of the largest earthquakes that should be labled (optional parameter)}
}
\description{
Building geom_timeline_label
}
\examples{
\dontrun{
ggplot(data = usa_data) +
geom_timeline(aes(x = DATE, color = DEATHS, size=EQ_PRIMARY), alpha=0.6) +
geom_timeline_label(aes(label = LOCATION_NAME, x = DATE, size=EQ_PRIMARY, n_max = 4)) +
labs(size = "Richter scale value", color = "# DEATHS") +
theme_classic() +
theme(legend.position="bottom")
}

}
